+++ ochrona/importer/import_wrapper.py
@@ -36,15 +36,15 @@ class SafeImport:
             packages = parsers.requirements.parse(package)
             if self._check_package(packages):
                 self._logger.info(
-                    f"No vulnerabilities found for {', '.join(packages)}, install proceeding."
+                    f"No vulnerabilities found for {', '.join([p.get('version') for p in packages])}, install proceeding."
                 )
                 self._install_file(package)
             else:
                 self._logger.error(
-                    f"Import of {', '.join(packages)} aborted due to detected vulnerabilities."
+                    f"Import of {', '.join([p.get('version') for p in packages])} aborted due to detected vulnerabilities."
                 )
         else:
-            if self._check_package(package):
+            if self._check_package({"version": package}):
                 self._logger.info(
                     f"No vulnerabilities found for {package}, install proceeding."
                 )
@@ -62,14 +62,14 @@ class SafeImport:
         :return: bool
         """
         vuln_results: DependencySet
-        if isinstance(package, str):
-            if any(spec in package for spec in INVALID_SPECIFIERS):
+        if isinstance(package, dict):
+            if any(spec in package.get("version") for spec in INVALID_SPECIFIERS):
                 raise OchronaImportException(
-                    f"An invalid specifier was found in {package}, either specify the package without a version or pin to a single version using `name==version`."
+                    f"An invalid specifier was found in {package.get('version')}, either specify the package without a version or pin to a single version using `name==version`."
                 )
-            parts = package.split(EQUALS_SPECIFIER)
+            parts = package.get("version").split(EQUALS_SPECIFIER)
             if len(parts) == 1:
-                package = self._get_most_recent_version(package=package)
+                package["version"] = self._get_most_recent_version(package=package.get("version"))
             vuln_results = resolve(dependencies=[package], logger=self._logger)
         elif isinstance(package, list):
             vuln_results = resolve(dependencies=package, logger=self._logger)